+++ examples/resources/meraki_appliance_vpn_bgp/resource.tf
@@ -5,17 +5,20 @@ resource "meraki_appliance_vpn_bgp" "example" {
   ibgp_hold_timer = 120
   neighbors = [
     {
-      allow_transit           = true
-      ebgp_hold_timer         = 180
-      ebgp_multihop           = 2
-      ip                      = "10.10.10.22"
-      next_hop_ip             = "1.2.3.4"
-      receive_limit           = 120
-      remote_as_number        = 64343
-      source_interface        = "wan1"
-      authentication_password = "abc123"
-      ipv6_address            = "2002::1234:abcd:ffff:c0a8:101"
-      ttl_security_enabled    = false
+      allow_transit            = true
+      ebgp_hold_timer          = 180
+      ebgp_multihop            = 2
+      ip                       = "10.10.10.22"
+      multi_exit_discriminator = 2
+      next_hop_ip              = "1.2.3.4"
+      receive_limit            = 120
+      remote_as_number         = 64343
+      source_interface         = "wan1"
+      weight                   = 10
+      authentication_password  = "abc123"
+      ipv6_address             = "2002::1234:abcd:ffff:c0a8:101"
+      ttl_security_enabled     = false
+      path_prepend             = [1]
     }
   ]
 }