+++ internal/provider/resource_meraki_switch_routing_multicast_rendezvous_points.go
@@ -100,6 +100,10 @@ func (r *SwitchRoutingMulticastRendezvousPointsResource) Schema(ctx context.Cont
 							MarkdownDescription: helpers.NewAttributeDescription("`Any`, or the IP address of a multicast group").String,
 							Required:            true,
 						},
+						"vrf_name": schema.StringAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("The name of the VRF").String,
+							Optional:            true,
+						},
 					},
 				},
 			},